pyproject.toml
[tool.poetry]
name = "django-square"
version = "0.1.0"
description = ""
authors = ["Jerome Leclanche <jerome@leclan.ch>"]
readme = "README.md"
packages = [{include = "django_square"}]

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2022.7.1"
proto-square-api = "^0.1.0"
django = "^4.1.7"
requests = "^2.28.2"
py-moneyed = "^3.0"
django-countries = "^7.5.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
